+++ design-system/page-header/examples/01-default.tsx
@@ -16,6 +16,7 @@ const PageHeaderDefaultExample = () => (
     <PageHeader breadcrumbs={breadcrumbs}>
       Title describing what content to expect on the page
     </PageHeader>
+    <div>This is some page content underneath.</div>
   </div>
 );