+++ editor/editor-markdown-transformer/src/index.ts
@@ -105,7 +105,6 @@ const mdToPmMapping = {
 
 const md = MarkdownIt('zero', {
   html: false,
-  linkify: true,
 });
 
 md.enable([