+++ navigation/atlassian-navigation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/atlassian-navigation
 
+## 1.1.2
+
+### Patch Changes
+
+- [`789bc630b95`](https://bitbucket.org/atlassian/atlassian-frontend/commits/789bc630b95) - Moved switcher test utils to private scope.
+
 ## 1.1.1
 
 ### Patch Changes